Allow user to click Send/Receive during or after txn and go back to send form
