+++ imports/startup/client/functions.js
@@ -167,6 +167,133 @@ getBalance = (getAddress, callBack) => {
   })
 }
 
+loadAddressTransactions = () => {
+  const thisTxs = LocalStore.get('address').state.transactions.reverse()
+
+  const request = {
+    tx: thisTxs,
+    grpc: findNodeData(DEFAULT_NODES, selectedNode()).grpc,
+  }
+
+  LocalStore.set('addressTransactions', [])
+  $('#loadingTransactions').show()
+  
+  Meteor.call('addressTransactions', request, (err, res) => {
+    if (err) {
+      LocalStore.set('addressTransactions', { error: err })
+    } else {
+      LocalStore.set('addressTransactions', res)
+      $('#loadingTransactions').hide()
+    }
+  })
+}
+
+getTokenBalances = (getAddress, callback) => {
+  const grpcEndpoint = findNodeData(DEFAULT_NODES, selectedNode()).grpc
+  const request = {
+    address: addressForAPI(getAddress),
+    grpc: grpcEndpoint,
+  }
+
+  Meteor.call('getAddress', request, (err, res) => {
+    if (err) {
+      // TODO - Error handling
+    } else {
+      if (res.state.address !== '') {
+        // Now for each res.state.token we find, go discover token name and symbol
+        for (let i in res.state.tokens) {
+          const tokenHash = i
+          const tokenBalance = res.state.tokens[i]
+
+          let thisToken = {}
+
+          const grpcEndpoint = findNodeData(DEFAULT_NODES, selectedNode()).grpc
+          const request = {
+            query: tokenHash,
+            grpc: grpcEndpoint,
+          }
+
+          Meteor.call('getTxnHash', request, (err, res) => {
+            if (err) {
+              // TODO - Error handling here
+              console.log('err:',err)
+            } else {
+              // Check if this is a token hash.
+              if (res.transaction.tx.transactionType !== "token") {
+                // TODO - Error handling here
+              } else {
+                let tokenDetails = res.transaction.tx.token
+
+                thisToken.hash = tokenHash
+                thisToken.name = bytesToString(tokenDetails.name)
+                thisToken.symbol = bytesToString(tokenDetails.symbol)
+                thisToken.balance = tokenBalance / SHOR_PER_QUANTA
+
+                tokensHeld.push(thisToken)
+
+                LocalStore.set('tokensHeld', tokensHeld)
+              }
+            }
+          })
+        }
+
+        callback()
+
+        // When done hide loading section
+        $('#loading').hide()
+      } else {
+        // Wallet not found, put together an empty response
+        callback()
+      }
+    }
+  })
+}
+
+updateBalanceField = () => {
+  const selectedType = document.getElementById('amountType').value
+
+  // Quanta Balances
+  if(selectedType == 'quanta') {
+    LocalStore.set('balanceAmount', LocalStore.get('transferFromBalance'))
+    LocalStore.set('balanceSymbol', 'Quanta')
+  } else {
+    // First extract the token Hash
+    tokenHash = selectedType.split('-')[1]
+
+    // Now calculate the token balance.
+    _.each(LocalStore.get('tokensHeld'), (token) => {
+      if(token.hash == tokenHash) {
+        LocalStore.set('balanceAmount', token.balance)
+        LocalStore.set('balanceSymbol', token.symbol)
+      }
+    })
+  }
+}
+
+refreshTransferPage = () => {
+  resetLocalStorageState()
+
+  // Wait for QRLLIB to load
+  waitForQRLLIB(function () {
+    // Get address balance
+    getBalance(getXMSSDetails().address, function() {
+      // Load Wallet Transactions
+      loadAddressTransactions()
+    })
+
+    // Get Tokens and Balances
+    getTokenBalances(getXMSSDetails().address, function() {
+      // Update balance field
+      updateBalanceField()
+      
+      $('#tokenBalancesLoading').hide()
+      
+      // Render dropdown
+      $('.ui.dropdown').dropdown()
+    })
+  })
+}
+
 // Reset wallet localstorage state
 resetLocalStorageState = () => {
   LocalStore.set('address', '')
@@ -179,4 +306,5 @@ resetLocalStorageState = () => {
   LocalStore.set('otsKeyEstimate', '')
   LocalStore.set('balanceAmount', '')
   LocalStore.set('balanceSymbol', '')
+  LocalStore.set('transactionConfirmed', 'false')
 }
