+++ imports/ui/pages/transfer/reload.html
@@ -1,0 +1,2 @@
+<template name="appReloadTransfer">
+</template>
