+++ imports/ui/pages/transfer/reload.js
@@ -1,0 +1,6 @@
+import './reload.html'
+
+Template.appReloadTransfer.onCreated(() => {
+  const path = FlowRouter.path('/transfer', {})
+  FlowRouter.go(path)
+})
